requirements/base.txt
Brotli==1.0.9
dj-database-url==0.5.0
Django==3.2.9
django-localflavor==3.1
mysqlclient==2.0.3
Pillow==8.4.0
whitenoise==5.3.0
